Move the last of classes: split tree into api/impl, moved Identifier and NodeOperations to api
