+++ src/main/java/me/zeroeightsix/fiber/api/annotation/exception/MalformedFieldException.java
@@ -2,9 +2,10 @@ package me.zeroeightsix.fiber.api.annotation.exception;
 
 import me.zeroeightsix.fiber.api.annotation.AnnotatedSettings;
 import me.zeroeightsix.fiber.api.exception.FiberException;
+import me.zeroeightsix.fiber.api.tree.ConfigBranch;
 
 /**
- * An exception thrown by {@link AnnotatedSettings AnnotatedSettings} during the conversion of a POJO to a {@link me.zeroeightsix.fiber.tree.ConfigBranch branch} when a field was not in the expected format.
+ * An exception thrown by {@link AnnotatedSettings AnnotatedSettings} during the conversion of a POJO to a {@link ConfigBranch branch} when a field was not in the expected format.
  */
 public class MalformedFieldException extends FiberException {
     public MalformedFieldException(String s) {
