+++ src/main/java/me/zeroeightsix/fiber/api/builder/ConfigTreeBuilder.java
@@ -5,7 +5,8 @@ import me.zeroeightsix.fiber.api.annotation.Setting;
 import me.zeroeightsix.fiber.api.annotation.Settings;
 import me.zeroeightsix.fiber.api.exception.FiberException;
 import me.zeroeightsix.fiber.api.exception.RuntimeFiberException;
-import me.zeroeightsix.fiber.tree.*;
+import me.zeroeightsix.fiber.api.tree.*;
+import me.zeroeightsix.fiber.impl.tree.ConfigBranchImpl;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
