+++ src/main/java/me/zeroeightsix/fiber/api/exception/FiberQueryException.java
@@ -1,9 +1,9 @@
 package me.zeroeightsix.fiber.api.exception;
 
-import me.zeroeightsix.fiber.tree.ConfigBranch;
-import me.zeroeightsix.fiber.tree.ConfigNode;
-import me.zeroeightsix.fiber.tree.ConfigQuery;
-import me.zeroeightsix.fiber.tree.ConfigTree;
+import me.zeroeightsix.fiber.api.tree.ConfigBranch;
+import me.zeroeightsix.fiber.api.tree.ConfigNode;
+import me.zeroeightsix.fiber.impl.tree.ConfigQuery;
+import me.zeroeightsix.fiber.api.tree.ConfigTree;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
