+++ src/main/java/me/zeroeightsix/fiber/api/schema/SchemaGenerator.java
@@ -5,14 +5,14 @@ import blue.endless.jankson.JsonElement;
 import blue.endless.jankson.JsonObject;
 import blue.endless.jankson.JsonPrimitive;
 import blue.endless.jankson.impl.MarshallerImpl;
-import me.zeroeightsix.fiber.Identifier;
+import me.zeroeightsix.fiber.api.Identifier;
 import me.zeroeightsix.fiber.api.builder.constraint.CompositeConstraintsBuilder;
 import me.zeroeightsix.fiber.api.constraint.Constraint;
 import me.zeroeightsix.fiber.impl.constraint.ValuedConstraint;
 import me.zeroeightsix.fiber.api.serialization.Marshaller;
-import me.zeroeightsix.fiber.tree.ConfigBranch;
-import me.zeroeightsix.fiber.tree.ConfigLeaf;
-import me.zeroeightsix.fiber.tree.ConfigTree;
+import me.zeroeightsix.fiber.api.tree.ConfigBranch;
+import me.zeroeightsix.fiber.api.tree.ConfigLeaf;
+import me.zeroeightsix.fiber.api.tree.ConfigTree;
 
 import javax.annotation.Nullable;
 import java.util.HashMap;
