+++ src/main/java/me/zeroeightsix/fiber/api/tree/Commentable.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.tree;
+package me.zeroeightsix.fiber.api.tree;
 
 /**
  * Capable of being marked with a comment.
