+++ src/main/java/me/zeroeightsix/fiber/api/tree/ConfigLeaf.java
@@ -1,8 +1,9 @@
-package me.zeroeightsix.fiber.tree;
+package me.zeroeightsix.fiber.api.tree;
 
 import me.zeroeightsix.fiber.api.builder.ConfigLeafBuilder;
 import me.zeroeightsix.fiber.api.constraint.Constraint;
 import me.zeroeightsix.fiber.impl.constraint.FinalConstraint;
+import me.zeroeightsix.fiber.impl.tree.ConfigNodeImpl;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
