+++ src/main/java/me/zeroeightsix/fiber/api/tree/HasValue.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.tree;
+package me.zeroeightsix.fiber.api.tree;
 
 /**
  * Indicates that this class holds some nullable value.
