+++ src/main/java/me/zeroeightsix/fiber/impl/tree/ConfigQuery.java
@@ -1,6 +1,7 @@
-package me.zeroeightsix.fiber.tree;
+package me.zeroeightsix.fiber.impl.tree;
 
 import me.zeroeightsix.fiber.api.exception.FiberQueryException;
+import me.zeroeightsix.fiber.api.tree.*;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
