+++ src/test/java/me/zeroeightsix/fiber/NodeOperationsTest.java
@@ -1,11 +1,12 @@
 package me.zeroeightsix.fiber;
 
+import me.zeroeightsix.fiber.api.NodeOperations;
 import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
 import me.zeroeightsix.fiber.api.builder.ConfigLeafBuilder;
-import me.zeroeightsix.fiber.tree.ConfigTree;
-import me.zeroeightsix.fiber.tree.ConfigLeaf;
-import me.zeroeightsix.fiber.tree.Property;
-import me.zeroeightsix.fiber.tree.ConfigNode;
+import me.zeroeightsix.fiber.api.tree.ConfigTree;
+import me.zeroeightsix.fiber.api.tree.ConfigLeaf;
+import me.zeroeightsix.fiber.api.tree.Property;
+import me.zeroeightsix.fiber.api.tree.ConfigNode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
