+++ src/test/java/me/zeroeightsix/fiber/tree/ConfigQueryTest.java
@@ -1,6 +1,9 @@
 package me.zeroeightsix.fiber.tree;
 
 import me.zeroeightsix.fiber.api.exception.FiberQueryException;
+import me.zeroeightsix.fiber.api.tree.ConfigLeaf;
+import me.zeroeightsix.fiber.api.tree.ConfigTree;
+import me.zeroeightsix.fiber.impl.tree.ConfigQuery;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.atomic.AtomicReference;
